--- src/components/ThemeButtons/colorButtons.module.scss
@@ -1,5 +1,0 @@
-.colorButtons {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
