+++ src/constants/theme.ts
@@ -1,0 +1,4 @@
+const COLOR_THEME = ['orange', 'green', 'blue', 'yellow', 'pink'];
+const BACKGROUND_THEME = ['light', 'dim', 'dark'];
+
+export { COLOR_THEME, BACKGROUND_THEME };
