+++ src/routes/Modal/ThemePicker/themePicker.module.scss
@@ -39,7 +39,7 @@
     line-height: 80px;
   }
 
-  & > div {
+  & > ul {
     height: 100px;
   }
