+++ src/routes/Portfolio/Aside/aside.module.scss
@@ -46,7 +46,7 @@
       padding: 0 20px;
       transform: translateX(-50%);
 
-      & > div:first-of-type {
+      & > ul:first-of-type {
         margin-bottom: 20px;
       }
     }
