+++ src/routes/Portfolio/portfolio.module.scss
@@ -62,7 +62,7 @@
     grid-template-columns: 300px 3fr;
     height: 100%;
     animation-name: slidein;
-    animation-duration: 2s;
+    animation-duration: 1s;
 
     .hidden {
       display: none;
@@ -110,7 +110,7 @@
 
 @keyframes slidein {
   from {
-    width: 300%;
+    width: 100%;
     margin-left: 100%;
   }
