+++ src/routes/ThemeSettingPage/themeSettingPage.module.scss
@@ -38,7 +38,7 @@
       line-height: 40px;
     }
 
-    & > div {
+    & > ul {
       height: 100px;
     }
